RavenDB-17760 : move tests to septate class, test bucket-stats with document extensions, cleanup
