+++ src/Raven.Server/Documents/DocumentsTransaction.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using Raven.Client.Documents.Changes;
 using Raven.Server.Documents.Changes;
 using Raven.Server.Documents.Replication.Incoming;
-using Raven.Server.Documents.Sharding;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
 using Voron.Impl;
