+++ test/SlowTests/Sharding/Cluster/ReshardingTests.cs
@@ -322,247 +322,8 @@ namespace SlowTests.Sharding.Cluster
                 }
             }
         }
-
-        [RavenFact(RavenTestCategory.Sharding)]
-        public async Task CanGetBucketStats()
-        {
-            var bucket = ShardHelper.GetBucket("users/1/$abc");
-
-            using (var store = Sharding.GetDocumentStore())
-            {
-                var before = DateTime.UtcNow;
-                using (var session = store.OpenAsyncSession())
-                {
-                    for (int i = 0; i < 10; i++)
-                    {
-                        await session.StoreAsync(new Raven.Tests.Core.Utils.Entities.User(), $"users/{i}/$abc");
-                    }
-
-                    await session.SaveChangesAsync();
-                }
-                var after = DateTime.UtcNow;
-
-                var db = await GetDocumentDatabaseInstanceFor(store, ShardHelper.ToShardName(store.Database, 1));
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatisticsFor(ctx, bucket);
-                    Assert.Equal(bucket, stats.Bucket);
-                    Assert.Equal(2811, stats.Size);
-                    Assert.Equal(10, stats.NumberOfItems);
-                    Assert.True(stats.LastModified > before);
-                    Assert.True(stats.LastModified < after);
-                }
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    before = DateTime.UtcNow;
-                    for (int i = 0; i < 10; i++)
-                    {
-                        if (i % 2 == 0)
-                            continue;
-                        var doc = await session.LoadAsync<Raven.Tests.Core.Utils.Entities.User>($"users/{i}/$abc");
-                        doc.Age = i * 8;
-                    }
-
-                    await session.SaveChangesAsync();
-                    after = DateTime.UtcNow;
-                }
-
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatisticsFor(ctx, bucket);
-                    Assert.Equal(bucket, stats.Bucket);
-                    Assert.Equal(2816, stats.Size);
-                    Assert.Equal(10, stats.NumberOfItems);
-                    Assert.True(stats.LastModified > before);
-                    Assert.True(stats.LastModified < after);
-                }
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    before = DateTime.UtcNow;
-                    for (int i = 0; i < 10; i++)
-                    {
-                        if (i % 2 != 0)
-                            continue;
-                        session.Delete($"users/{i}/$abc");
-                    }
-
-                    await session.SaveChangesAsync();
-                    after = DateTime.UtcNow;
-                }
-
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatisticsFor(ctx, bucket);
-                    Assert.Equal(bucket, stats.Bucket);
-                    Assert.Equal(1846, stats.Size);
-                    Assert.Equal(5, stats.NumberOfItems);
-                    Assert.True(stats.LastModified > before);
-                    Assert.True(stats.LastModified < after);
-                }
-            }
-        }
-
-        [RavenFact(RavenTestCategory.Sharding)]
-        public async Task CanGetBucketStats2()
-        {
-            var bucket1 = ShardHelper.GetBucket("users/1/$a");
-            var bucket2 = ShardHelper.GetBucket("users/1/$b");
-            var bucket3 = ShardHelper.GetBucket("users/1/$c");
-
-            var buckets = new Dictionary<int, (int NumOfDocs, int Size)>
-            {
-                [bucket1] = (10, 2771), 
-                [bucket2] = (20, 5611), 
-                [bucket3] = (30, 8431)
-            };
-
-            using (var store = Sharding.GetDocumentStore())
-            {
-                var before = DateTime.UtcNow;
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    var numOfDocs = buckets[bucket1].NumOfDocs;
-                    for (int i = 0; i < numOfDocs; i++)
-                    {
-                        await session.StoreAsync(new Raven.Tests.Core.Utils.Entities.User(), $"users/{i}/$a");
-                    }
-                    await session.SaveChangesAsync();
-                }
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    var numOfDocs = buckets[bucket2].NumOfDocs;
-                    for (int i = 0; i < numOfDocs; i++)
-                    {
-                        await session.StoreAsync(new Raven.Tests.Core.Utils.Entities.User
-                        {
-                            Name = "b"
-                        }, $"users/{i}/$b");
-                    }
-                    await session.SaveChangesAsync();
-                }
-
-                using (var session = store.OpenAsyncSession())
-                {
-                    var numOfDocs = buckets[bucket3].NumOfDocs;
-                    for (int i = 0; i < numOfDocs; i++)
-                    {
-                        await session.StoreAsync(new Raven.Tests.Core.Utils.Entities.User
-                        {
-                            Name = "c",
-                            Age = i
-                        }, $"users/{i}/$c");
-                    }
-
-                    await session.SaveChangesAsync();
-                }
-
-                var after = DateTime.UtcNow;
-                var record = await store.Maintenance.Server.SendAsync(new GetDatabaseRecordOperation(store.Database));
-                var shard = ShardHelper.GetShardNumber(record.Sharding.BucketRanges, bucket1);
-
-                var db = await GetDocumentDatabaseInstanceFor(store, ShardHelper.ToShardName(store.Database, shard));
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatisticsFor(ctx, bucket1);
-                    var expected = buckets[bucket1];
-
-                    Assert.Equal(bucket1, stats.Bucket);
-                    Assert.Equal(expected.Size, stats.Size);
-                    Assert.Equal(expected.NumOfDocs, stats.NumberOfItems);
-                    Assert.True(stats.LastModified > before);
-                    Assert.True(stats.LastModified < after);
-                }
-
-                shard = ShardHelper.GetShardNumber(record.Sharding.BucketRanges, bucket2);
-                db = await GetDocumentDatabaseInstanceFor(store, ShardHelper.ToShardName(store.Database, shard));
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatisticsFor(ctx, bucket2);
-                    var expected = buckets[bucket2];
-
-                    Assert.Equal(bucket2, stats.Bucket);
-                    Assert.Equal(expected.Size, stats.Size);
-                    Assert.Equal(expected.NumOfDocs, stats.NumberOfItems);
-                    Assert.True(stats.LastModified > before);
-                    Assert.True(stats.LastModified < after);
-                }
-
-                shard = ShardHelper.GetShardNumber(record.Sharding.BucketRanges, bucket3);
-                db = await GetDocumentDatabaseInstanceFor(store, ShardHelper.ToShardName(store.Database, shard));
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatisticsFor(ctx, bucket3);
-                    var expected = buckets[bucket3];
-
-                    Assert.Equal(bucket3, stats.Bucket);
-                    Assert.Equal(expected.Size, stats.Size);
-                    Assert.Equal(expected.NumOfDocs, stats.NumberOfItems);
-                    Assert.True(stats.LastModified > before);
-                    Assert.True(stats.LastModified < after);
-                }
-            }
-        }
-
-        [RavenFact(RavenTestCategory.Sharding)]
-        public async Task CanGetBucketStatsForRange()
-        {
-            using (var store = Sharding.GetDocumentStore())
-            {
-                var numOfDocsPerBucket = 10;
-                var bucketsInShard0 = new HashSet<int>();
-
-                var record = await store.Maintenance.Server.SendAsync(new GetDatabaseRecordOperation(store.Database));
-
-                for (int i = 1; i <= 100; i++)
-                {
-                    var suffix = i.ToString();
-                    var bucket = ShardHelper.GetBucket($"users/1/${suffix}");
-                    var shard = ShardHelper.GetShardNumber(record.Sharding.BucketRanges, bucket);
-
-                    if (shard != 0)
-                        continue;
-
-                    if (bucketsInShard0.Add(bucket) == false)
-                        continue;
-
-                    using (var session = store.OpenAsyncSession())
-                    {
-                        for (int j = 0; j < numOfDocsPerBucket; j++)
-                        {
-                            await session.StoreAsync(new Raven.Tests.Core.Utils.Entities.User(), $"users/{j}/${suffix}");
-                        }
-
-                        await session.SaveChangesAsync();
                     }
                 }
-
-                var db = await GetDocumentDatabaseInstanceFor(store, ShardHelper.ToShardName(store.Database, shard: 0));
-                using (db.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext ctx))
-                using (ctx.OpenReadTransaction())
-                {
-                    var stats = ShardedDocumentsStorage.GetBucketStatistics(ctx, fromBucket: 0, toBucket: int.MaxValue).ToList();
-                    Assert.Equal(bucketsInShard0.Count, stats.Count);
-
-                    foreach (var bucketStats in stats)
-                    {
-                        Assert.Equal(numOfDocsPerBucket, bucketStats.NumberOfItems);
-                        Assert.True(bucketStats.Size > 0);
-                    }
-                }
-            }
-        }
-    }
-}
 
         [RavenFact(RavenTestCategory.Sharding)]
         public async Task BucketDeletionShouldCreateArtificialTombstones()
